fix: 온보딩 Fragment 내 날짜 포맷 함수 Util 객체에서 관리
